Revert removing ChaikaCore.getServerURL refs #234
I decided to replace the usage of ChaikaCore with that of new developed
modules and keep untouched any methods and classes in ChaikaCore.js,
because changing or removing them from the file breaks the existing
related add-ons like LogSearch2ch.

It's better to notify developers that it's now deprecated and will be
removed in the future version of chaika.