Stop trying to open a file directly in ChaikaIO#reveal. refs #283
I found opening a file via nsIFile#launch is not a good idea because
chaika cannot know what application is associated with a specified
file, which may lead to open the file in an inappropriate and
unsuitable application.

Revealing the file in user's system filer, i.e., in Finder, File
Explorer or something like them, is far more better than opening the
file in meaningless software.