Use border to highlight and adjust margin & padding of post.
1. I learned Chameleon orange, or `--in-content-border-highlight` is
originally intended to being used as a border color and not as a
background color. That's why the orange color is so deep.

Suggested by: http://anago.2ch.net/test/read.cgi/software/1427337153/928

2. Previously both margins and paddings of `.resBody` and
`.resContainer` are used to adjust the space between each posts. This
is very confusing and hard to understand the code meaning.
This commit resolve it by clarifying each roles of the two element:
- `.resContainer` to set margins between posts.
- `.resBody` to set margins around a message body within a post.